Update and rename lapi.c to lapi.c
